+++ src/content/app/tools/blast/views/blast-submission-results/BlastSubmissionResults.scss
@@ -6,144 +6,3 @@ $double_standard_gutter: calc($standard-gutter * 2);
   max-width: 1860px;
   padding: 0 $standard-gutter $global-padding-bottom $double_standard_gutter;
 }
-
-.sequenceBoxWrapper {
-  padding: 24px 20px 0px 20px;
-  border: 1px solid $grey;
-}
-
-.sequenceBoxWrapper + .sequenceBoxWrapper {
-  border-top: none;
-}
-
-.sequenceHeader {
-  display: flex;
-  flex-direction: column;
-
-  .content {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-}
-
-.submissionParameters {
-  display: grid;
-  grid-template-areas:
-    'main-top aside-top'
-    'main-bottom aside-bottom';
-  grid-template-columns: min-content 1fr;
-  grid-template-rows: 20px auto;
-  align-items: start;
-  grid-row-gap: 17px;
-  column-gap: 100px;
-
-  .copySequence {
-    justify-self: flex-end;
-    grid-area: main-top;
-  }
-
-  .sequence {
-    width: calc(60ch + 14px);
-    grid-area: main-bottom;
-    margin-top: 10px;
-    min-height: 170px;
-    max-height: 430px;
-    font-family: $font-family-monospace;
-    color: $dark-grey;
-    overflow-y: auto;
-    overflow-x: hidden;
-    scrollbar-gutter: auto;
-  }
-
-  .parameters {
-    grid-area: aside-bottom;
-
-    tr {
-      height: 30px;
-
-      td:first-of-type {
-        text-align: right;
-        font-weight: $light;
-        padding-right: 10px;
-      }
-    }
-
-    .parameterCheckbox {
-      height: 12px;
-      width: 12px;
-      background-color: $green;
-      border: 1px solid $grey;
-      display: inline-block;
-      margin-right: 10px;
-      vertical-align: middle;
-    }
-
-    .parameterCheckbox + span {
-      font-weight: $light;
-    }
-  }
-}
-
-.resultsSummaryRow {
-  display: grid;
-  grid-template-columns: [sequence-label] 11ch [hits-diagram] minmax(
-      48%,
-      1050px
-    ) [species] minmax(20%, 1fr) 10px;
-  column-gap: 20px;
-  padding-bottom: 20px;
-}
-
-.hitLabel {
-  text-align: right;
-  padding-right: 20px;
-}
-
-.summaryPlot {
-  grid-column: hits-diagram;
-}
-
-.againstText {
-  font-weight: $light;
-}
-
-.speciesInfo span {
-  padding-right: 20px;
-}
-
-.submissionIdLabel {
-  font-weight: $light;
-  margin-right: 1ch;
-}
-
-.editSubmission {
-  color: $blue;
-  cursor: pointer;
-  margin: 0 32px;
-}
-
-.timeStamp {
-  white-space: nowrap;
-}
-
-.timeZone {
-  font-weight: $light;
-  margin-left: 1ch;
-}
-
-.controlButtons {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  column-gap: 40px;
-  width: 268px;
-}
-
-.showHideWrapper {
-  --show-hide-chevron-height: 8px;
-}
-
-.rulerWrapper {
-  padding-bottom: 10px;
-}
