+++ src/content/app/tools/blast/views/blast-submission-results/components/sequence-box/SequenceBox.scss
@@ -1,0 +1,46 @@
+@import 'src/styles/common';
+
+.resultsSummaryRow {
+  display: grid;
+  grid-template-columns:
+    [sequence-label] 11ch [hits-diagram] minmax(48%, 1050px)
+    [species] minmax(20%, 1fr) 10px;
+  column-gap: 20px;
+  padding-bottom: 20px;
+}
+
+.rulerWrapper {
+  padding-bottom: 10px;
+}
+
+.sequenceBoxWrapper {
+  padding: 24px 20px 0px 20px;
+  border: 1px solid $grey;
+}
+
+.sequenceBoxWrapper + .sequenceBoxWrapper {
+  border-top: none;
+}
+
+.sequenceHeader {
+  display: flex;
+  flex-direction: column;
+
+  .content {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+}
+
+.againstText {
+  font-weight: $light;
+}
+
+.showHideWrapper {
+  --show-hide-chevron-height: 8px;
+}
+
+.summaryPlot {
+  grid-column: hits-diagram;
+}
