+++ src/content/app/tools/blast/views/blast-submission-results/components/single-blast-job-result/SingleBlastJobResult.scss
@@ -1,0 +1,21 @@
+.resultsSummaryRow {
+  display: grid;
+  grid-template-columns:
+    [sequence-label] 11ch [hits-diagram] minmax(48%, 1050px)
+    [species] minmax(20%, 1fr) 10px;
+  column-gap: 20px;
+  padding-bottom: 20px;
+}
+
+.hitLabel {
+  text-align: right;
+  padding-right: 20px;
+}
+
+.summaryPlot {
+  grid-column: hits-diagram;
+}
+
+.speciesInfo span {
+  padding-right: 20px;
+}
